うっかり閉じるボタンを押しちゃったときにアラート出す機能
Fix #699
