+++ src/js/Components/Views/Options/Settings/Uncategorized/index.js
@@ -14,51 +14,13 @@ export default class UncategorizedSettings extends Component {
             <div>
               <h1 style={this.props.styles.title}><Settings /> 未分類の設定置き場</h1>
               <Description>わりと設定項目が細かいし隅々まで散らばってるので、セクションつくるほど類似した設定がない設定はここにつっこむ的な。ゲーム内ボタン表示はとりあえずAPPモードのみ対応です。好評だったらWHITEモードにも追加します。</Description>
-              <DamageSnapshotSetting />
               <DashboardLayoutSetting />
-              <UseInAppActionButtonsSetting />
               <PopupBackgroundImageSetting />
             </div>
         );
     }
     static propTypes = {
         styles: PropTypes.object.isRequired,
-    }
-}
-
-// 未分類クラス！！
-class DamageSnapshotSetting extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            model: Config.find("damagesnapshot-window")
-        };
-    }
-    render() {
-        return (
-            <Table selectable={false}>
-              <TableBody displayRowCheckbox={false}>
-                <TableRow>
-                  <TableRowColumn>
-                    大破進撃防止窓
-                  </TableRowColumn>
-                  <TableRowColumn>
-                    <SelectField value={this.state.model.value} onChange={this.onChange.bind(this)}>
-                      <MenuItem value="disabled" primaryText="使用しない"   />
-                      <MenuItem value="inwindow" primaryText="ゲーム画面左上"/>
-                      <MenuItem value="separate" primaryText="別窓表示"     />
-                    </SelectField>
-                  </TableRowColumn>
-                </TableRow>
-              </TableBody>
-            </Table>
-        );
-    }
-    onChange(ev, i, value) {
-        let model = this.state.model;
-        model.value = value;
-        model.save();
-        this.setState({model});
     }
 }
 
@@ -96,36 +58,7 @@ class DashboardLayoutSetting extends Component {
     }
 }
 
-class UseInAppActionButtonsSetting extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {
-            model: Config.find("use-inapp-action-buttons"),
-        };
-    }
-    render() {
-        return (
-          <Table selectable={false}>
-            <TableBody displayRowCheckbox={false}>
-              <TableRow>
-                <TableRowColumn>
-                  ゲーム画面にミュートなどのボタンを表示する
-                </TableRowColumn>
-                <TableRowColumn>
-                  <Toggle toggled={this.state.model.value} onToggle={this.onToggle.bind(this)}/>
-                </TableRowColumn>
-              </TableRow>
-            </TableBody>
-          </Table>
-        );
-    }
-    onToggle(ev, value) {
-        let model = this.state.model;
-        model.value = value;
-        model.save();
-        this.setState({model});
-    }
-}
+
 
 import {grey400}  from "material-ui/styles/colors";
 import Clear      from "material-ui/svg-icons/content/clear";
