+++ src/js/Components/Views/Options/index.js
@@ -10,6 +10,7 @@ import ScreenShotSettingsView   from "./Settings/ScreenShots";
 import SyncSettingsView         from "./Settings/Syncs";
 import TimerSettingsView        from "./Settings/Timer";
 import TwitterSettingsView      from "./Settings/Twitter";
+import InAppSettingsView        from "./Settings/InApp";
 import UncategorizedSettings    from "./Settings/Uncategorized";
 import Announcement   from "./Announcement";
 import WinconfigsView from "./Winconfigs";
@@ -51,6 +52,7 @@ export default class OptionsView extends Component {
             <TimerSettingsView        styles={styles} />
             <SyncSettingsView         styles={styles} />
             <TwitterSettingsView      styles={styles} />
+            <InAppSettingsView        styles={styles} />
             <UncategorizedSettings    styles={styles} />
             <WinconfigsView           styles={styles} />
             <ReferencesView           styles={styles} />
